+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [0.29.2] - 2026-06-02
+### Added
+- add browser agent
 
 ### Fixed
 - fix age formatting